Move old implementations and benchmark into subdirectory
